+++ money-maven/src/main/java/lt/techin/moneymaven/service/IncomeService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import lt.techin.moneymaven.dto.IncomeDto;
 import lt.techin.moneymaven.exception.IncomeNotFoundException;
 import lt.techin.moneymaven.exception.NoEntriesFoundException;
+import lt.techin.moneymaven.exception.UnauthorizedAccessException;
 import lt.techin.moneymaven.exception.UserNotFoundException;
 import lt.techin.moneymaven.model.Income;
 import lt.techin.moneymaven.model.User;
@@ -27,9 +28,9 @@ public class IncomeService {
 	@Autowired
 	private ModelMapper modelMapper;
 	
-	public Page<IncomeDto >getIncomes(Pageable pageable){
+	public Page<IncomeDto >getIncomes(Pageable pageable, Integer userId){
 		try {
-			Page<Income> incomes = incomeRepository.findAll(pageable);
+			Page<Income> incomes = incomeRepository.findAllByUser_userId(pageable, userId);
 			if (incomes.isEmpty()) {
 				throw new NoEntriesFoundException("incomes");
 			}
@@ -42,23 +43,30 @@ public class IncomeService {
 		
 	}
 	
-	public IncomeDto getIncomeById(Integer id) {
+	public IncomeDto getIncomeById(Integer id, Integer userId) {
 		try {			
 			Income income = incomeRepository.findById(id)
 					.orElseThrow(() -> new IncomeNotFoundException("Income Id", id));
+			
+			if (!income.getUser().getUserId().equals(userId)) {
+	            throw new UnauthorizedAccessException("Income", id, userId);
+	        }
+			
 			return modelMapper.map(income,  IncomeDto.class);
 		} catch (IncomeNotFoundException e) {
+			throw e;
+		} catch (UnauthorizedAccessException e){
 			throw e;
 		} catch (Exception e) {
 			throw new RuntimeException("Error while getting income", e);
 		}
 	}
 	
-	public IncomeDto createIncome(IncomeDto incomeDto) {
+	public IncomeDto createIncome(IncomeDto incomeDto, Integer userId) {
 		try {
 			Income income = modelMapper.map(incomeDto, Income.class);
-			User user = userRepository.findById(1)
-					.orElseThrow(() -> new UserNotFoundException("User Id", 1));
+			User user = userRepository.findById(userId)
+					.orElseThrow(() -> new UserNotFoundException("User Id", userId));
 			income.setUser(user);
 			Income savedIncome = incomeRepository.save(income);
 			return modelMapper.map(savedIncome, IncomeDto.class);
@@ -69,26 +77,40 @@ public class IncomeService {
 		}
 	}
 	
-	public IncomeDto updateIncome(Integer id, IncomeDto incomeDto) {
+	public IncomeDto updateIncome(Integer id, IncomeDto incomeDto, Integer userId) {
 		try {
 			Income existingIncome = incomeRepository.findById(id)
 					.orElseThrow(() -> new IncomeNotFoundException("Income Id", id));
+			
+			if (!existingIncome.getUser().getUserId().equals(userId)) {
+	            throw new UnauthorizedAccessException("Income", id, userId);
+	        }
+			
 			modelMapper.map(incomeDto, existingIncome);
 			Income savedIncome = incomeRepository.save(existingIncome);
 			return modelMapper.map(savedIncome, IncomeDto.class);
 		} catch (IncomeNotFoundException e) {
+			throw e;
+		} catch (UnauthorizedAccessException e){
 			throw e;
 		} catch (Exception e) {
 			throw new RuntimeException("Error while updating income", e);
 		}
 	}
 	
-	public void deleteIncome(Integer id) {
+	public void deleteIncome(Integer id, Integer userId) {
 		try {
 			Income income = incomeRepository.findById(id)
 			.orElseThrow(() -> new IncomeNotFoundException("Income_id", id));
+			
+			if (!income.getUser().getUserId().equals(userId)) {
+	            throw new UnauthorizedAccessException("Income", id, userId);
+	        }
+			
 			incomeRepository.delete(income);
 		} catch (IncomeNotFoundException e) {
+			throw e;
+		} catch (UnauthorizedAccessException e){
 			throw e;
 		} catch (Exception e) {
 			throw new RuntimeException("Error while deleting income", e);
